[squashme] refactor
